requirements.txt
requests==2.31.0
python-dotenv==1.0.0
streamlit==1.31.0
pydantic==2.5.2
pandas==2.1.4
numpy==1.24.3
plotly==5.18.0
python-dateutil==2.8.2
beautifulsoup4==4.12.2
lxml==4.9.3
aiohttp==3.9.1
cachetools==5.3.2
tushare==1.2.89 